server/requirements.txt
loguru==0.5.3
pika==1.2.0
grpcio 
grpcio-tools